requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
faiss-gpu==1.7.2
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.23.1
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightly==1.5.4
lightly-utils==0.0.2
lightning-utilities==0.11.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pydantic==1.10.15
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.5.10
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.2
